third_party/blink/web_tests/virtual/fenced-frame-shadow-dom/external/wpt/html/cross-origin-embedder-policy/anonymous-iframe/fenced-frame.tentative.https.window-expected.txt

This is a testharness.js-based test.
FAIL FencedFrame within an AnonymousIframe is not anonymous assert_equals: Check window.anonymous in FencedFrame expected "false" but got "true"
Harness: the test ran to completion.
